LogicRays/LogicraysTask/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.13.1
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
